Moved mpc related stuff in util.rs to mpc.rs
