+++ integration-tests/src/lib.rs
@@ -10,6 +10,7 @@ use near_workspaces::{
 use crate::env::containers;
 
 pub mod env;
+pub mod mpc;
 pub mod sandbox;
 pub mod util;
